Hard challenges in naive
After implementing the "saturable", "linear speed, linear gamma"
controller, it became clear that it would not work. Right now, I am
pretty sure the "linear speed, linear gamma" maneuver is impossible
unless the bounds on the angle of attack are very large so as to
not saturate.

For a moment, I even thought this case would be unfeasible.
Fortunately, though, there has been a significant development in
probRock, but this goes into another branch.

Right now, naive is still not capable of running Miele's (2003)
monostage case. I am working in a theoretical front to allow put-
ting better bounds on the errors.