Arc removal made it work!!
After all the failures in the naive branch, I decided to try a some-
what desperate approach, which was to increase the rocket mass AFTER
itsme delivered the solution, in order to meet the absurd Thrust-to-
weight ratio of 1.3 by Miele. It did not work at first, even alter-
nating mass increases and restoration cycles.

It then occurred to me that the solution in the first two arcs was
barely altered, but the one on the third, largest arc was almost
unrecognizable after the restorations. I decided to "remove" the
first two arcs, putting just one target height at 250 km, hence
making two arcs much more similar in duration. It worked! Not only
it survived the harsh mass increase after itsme (up to TWR=1.28, due
to an error in the code...), it converged in MSGRA with tolQ=1e-4.

Next steps are to try to remove the extra arc completely (does the
code allow that? I don't know!) and adjust also the wing loading.