Added accel limit to probLand... same problems
After all the difficulties with the acceleration limitation in prob-
Rock, I decided to turn to probLand instead. As expected, there were
a lot of problems as well. Tuning the Kpf constant seems to be a
very complicated issue: too low, the limitations are solemnly igno-
red; too high and the program just fails to converge. I don't have
many more ideas to try...