Changing > for >= in probLand did not work as well
On March 25th I tried changing the acceleration limitation conditi-
on for "activating" the additional cost term. It also did not work.
I began to question if these acceleration limitation problems even
have a true mathematical solution.

After a lot of mathematical work, I discovered that there is a
very big chance that probLand admits a true unique mathematical
solution for any Kpf. It would be good to try and code the
"internal" (mathematical) functions that guide the solution to see
if the approximation yielded by the program is at least close to
the true optimal solution. The best thing would be to try the same
method on a even simpler problem, like probCart, in order to
actually get an explicit analytical theoretical solution to compare
to the one given by the program.

For some reason it took me more than a month to commit this...