Added speed limitation to probCart, like probRock and probLand
Still trying to figure out why the acceleration limitation is so
hard to work with in both probLand and probRock, I decided to add
a speed limitation (with a Kpf * (v-vLim)**2 term, similar to the
former problems), and it works. In fact, since probCart is so much
simpler than rocket-related problems, I could also deduce the true
analytical solution to this modified version of probCart and wrote
a module that gives this solution.

(Actually, the analytical solution is not fully explicit. The time
equation yields a very ugly polynomial to solve for the costate of
the position, so it was better to write the solution in terms of
this costate and adjust it until the residual is small enough. Just
a simple disclaimer.)

It turns out the true solution in this limitated problem ends up
violating the limitation, always. In order to minimize the viola-
tion, one must set the Kpf parameter to a high value, which takes
a huge toll on (M)SGRA. So, it could be that MSGRA is just too
sensitive in probLand with respect to Kpf. At this point I am even
considering introducing a dynamic change of Kpf as the algorithm
converges. Still, it would be useful to have some bounds on the
ammount of acceleration violation as a function of Kpf parameter.

I have already worked out an analytical structure for the optimal
solution in probLand, but still have not coded it. It is much
more complicated than its probCart sibling, and will definitely be
much less analytical. Let's do it!