Analytical optimal solutions in probLand as well!
After a truly obscene ammout of time, I finally managed to get the
module analytiLand.py, which is analytiCart.py's cousin, fully
coded and running. It was a huge effort, in terms of both mathema-
tical and programming-related work.

In the end, I wrote a module that performs a hybrid between adjus-
ting constants searching for an optimal solution with a "fixed
shape" and actually solving the optimality ODEs. In probLand, the
optimal solution is always to do a two-segment descent: first a
free fall, then a powered descent, but the acceleration does not
stay at a constant value.

Now I should be ready to run probLand in MSGRA again, with several
values of Kpf, eventually giving away the optimal solution (obtai-
ned in analytiLand) as initial guess to MSGRA.