Configuration parameters to configuration file
OK, so now monthly commits are a thing.

It was actually rather easy to get the configuration parameters
put of the interf.py source code and put them in the .its file. I
really should have done this previously.

I am going to take some time and study the flow of the code to see
if I can improve this "parameter loading" section. Only then I'll
proceed to alter main.py so it can be imported by another module,
which will be used to do "batch runs", changing parameters such as
Kpf, GSS_PlimCte, etc.