Compatibilization of .its files
Some commits ago (954362407a3219cec5280efe3de4f5a4704a8032), I in-
cluded the configuration parameters into the .its file, but I did
it rather poorly. Every configuration file had to be changed in
order to include the [settings] section, which was now done.

I also did what I should have done previously and put a simple
test before trying to load the [settings] section. So now, if
there is no such section on the configuration file, the program
just defaults to the parameters hardcoded in interf.py (as it was
before).