Added file for batch runs
After a long time, I finally created the file for running main in
a sort of batch mode.

For now, there are two main uses that I have in mind. The first one
is 'explicit mode' in which the user explicitly lists the cases to
be run (possibly from different problems, in fact). This would be
great for tests, so the user can run all the problems to check for
compatibility breaks, for example.

The second one would be basically for automating parameter seeking.
I will start with the first one.