Batch mode is working!
Ok, so batch mode is actually working!

I started with the 'explicit' mode, and seems to work well, gene-
rating subfolders for each individual run inside the main batch
folder. It also handles eventual errors in individual runs without
interrupting the main loop.

I do need to keep changing the code so that the used does not have
do manually hit enter all the time. This should not be so hard.