Batch mode overrides user inputs and sounds
Now, when running batch mode, the user is not prompted to enter
anything. This was achieved by changing all instances of the
input() command to a fancy prom() version implemented in the
logger object.

Also, when running batch mode, the sound is supressed. At least
I could not find any other occurrence of print("\a")...