'Variations' mode for batch runs - parsing ok
I am working now on the 'variations' mode for batch runs, so the
user can produce variations on top of a 'base run'.

Right now, the code only parses the input from the user. The
"language" for specifying the variations seems simple enough.
I still have to develop the part that writes the .its files for
each run, and I intend to do it now.