'Variations' mode for batch runs is working!
I have finished the part that writes the .its files for each run
in the 'variations' mode, and I have been testing it with probLand.
Everything seems to be working properly.

Using this mode, I have already figured out that increasing N and
decreasing GSS_PLimCte, the program runs probLand with values of
PFtol for which there was no convergence before. All experiments in
the current version of land-KpfSweep.bat have finished well.

The next big inovation is to allow the initial guess of each run in
the experiment to be the final solution of the last. This will be
specially useful when the experiment itself is to lower
sequentially the tolQ parameter.