New plot added: Q history vs I history
Just added a new plot, that shows the Q convergence history, but
with I history (actually, I reduction history) as the x axis. This
is possible because I is (or should be!) strictly descending.

This little plot is actually very interesting because it provides
great insight into the selection of the tolQ parameter. If the I
reduction is negligible from a given order of Q magnitude onwards,
then it makes no sense to waste computational effort reducing Q
further (it that is even possible).

In fact, by running all the problems (except probRock, of course),
the same Q vs I reduction "shape" is essentially the same: at first
there is essentially no Q reduction as I decreases, and then there
is a steep Q descent with almost no I variation. I might be taking
early conclusions here, but I do think that the regularity of this
shape, which I refer to as the "knee curve", can be used to stop
the program earlier, without losing significant "optimization per-
formance".