Added file for indication of the status of the run
I've added a bit of code that creates simple text files reporting
on the status of the run (initializing, in progress, stopped by
user, error, etc). The idea is to provide a summary for the
anxious user so he/she can know where each run is without having
to open the log file.

My initial idea was to provide the (gradient) iteration in which
the program incurred in error or was stopped by the user, but in
these circunstances it seems the sol object is not properly
rewritten in the main module excecution, so the variable NIterGrad
is never updated from 1. If anyone knows a simple way around that,
feel free to do it.