Omission of unnecessary variations (issue #91)
I have implemented a rather simple approach to solve issue #91.
Basically, it is possible to just avoid propagating the solutions
in the LMPBVP corresponding to fixed-state boundary conditions and
remove the corresponding equations from the final linear system
to keep it from becoming over-determined.

The code of each problem must be changed for this to be applicable,
naturally, since the judgement of which variations are necessary or
not is not easily implementable. The alteration in each problem is
simply the addition of a few lines of code at the initialization
section.

I have implemented this in a way such that MSGRA still works
regardless if each problem has been altered or not. Hence, I have
only changed probCart, probLand, and of course, probRock.

The results in each iteration are identical (as they should be!),
and the performance gains seem promising.