Slight improvement in calcStepGrad performance
I got some basic improvement for the performance in calcStepGrad.
The main changes are to detect if the objective is strictly descen-
ding with the step (in this case, there is no need for detailed
step search) and the beginning of an improvement in findStepLim.

The idea is to keep improving calcStepGrad, but soon it won't be
the bottleneck anymore.

I've also modified probBrac for omission of unnecessary variations.