Huge changes to calcStepGrad
I have performed a huge change to calcStepGrad. There is no more
fitting of parabolas to the objective function, and instead a
"trissection" - a kind of golden section search that halves the
intervals instead of using the golden ratio.

The restricions such as P <= lim P, Obj <= Obj0 and the pi res-
trictions are handled by monitoring both the smallest value
that violates the restriction and the largest value that does
not violate it.

I believe the algorithm is now capable of achieving the minimum
value of objective with fewer objective function evaluations. I
still need to clean up the code and add the parameters to the
external configuration file.