Gradient test at step limit and duplicate calcQ removed
Further improving the calcStepGrad function, I added a test
for the gradient of the objective with respect to the step
after the detection of the step limit. If the gradient is
negative then the best step is precisely the one at the limit
since the actual min obj is after the limit.

I also removed the duplicate calculation of calcQ from the
gradRestCycl, which really could have been done earlier.

Testing the program with probRock, however, I could not even
get the end of the execution. It is clear once more that the
rocket problem is of a different nature altogether. New
improvements are still necessary.