Post-processing in batches
I added a little infrastructure to batch.py in order to properly
assess the effectiveness of the changes to the step searching
algorithm.

Now batch runs are followed by a post processing step in which
data can be retrieved from the sol object and displayed in a .xlsx
spreadsheet afterwards. New fields can easily be added for other
studies as well, of course.