Traceback output in log and solution loading
I've decided to work in some old problems. The first one is
writing the traceback message in the log file to allow for easier
debugging. The second one is to set the logger to manual before
saving the solution; this allows for loading the solution manually
to access the attributes and methods (which is must more likely).

I have also added a new way for leaving the grad-rest cycle if
there are too many iterations. This will be useful for running
problems like probLand and probRock in batch mode during a PFtol
search, for example.