Previous step information and spreadsheet writing in calcStepGrad
I had to decide which information from the previous step I would
use for the next calcStepGrad upgrade, so I added an attribute
with most of the information of the previous step, to be passed to
the next one.

I also added a "spreadsheet mode" to write all the tried values of
step and the respective properties in every objective evaluation.