Bug fix for spreadsheet writing and minor boost to printL
The spreadsheet-writing code had a bug due to the use of the folder
name on the spreadsheet name. This only revealed itself when tested
with batch.py. It was much harder to figure out than to fix...

I also took some time to take the string comparison in every single
printL command. After all the idea is to optimize run time, right?