No more duplicate sorted lists in grad_sgra
I have refactored the testDecr() method in order to stop using two
lists with the steps in ascending order and the objectives in the
same order. Basically only a list with the indexes is used now,
which makes the code more memory efficient.

There was also a tiny mistake in tryStep(), and the trios were set
before checking the validity of the step (and the eventual increase
in objective that follows when P>limP). This made the obj behavior
in probBrac a lot less smooth, which led to poorer performance. This
confused me for a long time, but what is wrong is wrong, even if it
makes some instances of some problems converge faster...