Optimizing atmospheric density calculation
After running a profiled probRock run, I had noticed that the atmos-
pheric density calculation took a long time to run. So, I decided to
take a look into atmosphere.py and, despite of issue #39, I have
written a new atmospheric density function which is much more effi-
cient than the older version (around 50% of the run time only).

We still have to enhance this model, though, as stated in issue #99.
This calls for a PR to master branch. Because of this, I have added
the .bat and .its modifications (which have nothing to do with the
current subject, but do not make that much of a difference).