More little advances for calcStepGrad
A lot of little advances:

- re-organizing the __init__() method
- added a history for the status (Good, Bad) of each step, for
easier decision
- added a "fast forward rate", replacing the "decade" in adjsTrios()
and in section 2.0
- added a safety saturation for alfaLimP > pi limit
- changed the check() method for outputting the violation, giving
more information to the upper method
- changed the check() method for not always artificially increasing
the value of objective, which messed up the minObj search
- changed tryStep() method to allow the monotonicity check if the
only violation was Obj>Obj0, this allows for stopping the P limit
seach earlier
- removed the trio assembly on 2.0.3, since it is already being done
automatically in tryStep()
- changed the condition for the gradient test in 2.1

In the end, the performance was a little worse in probBrac and
probLand, and better in 10_2. These are the results for testAll:

Problem name  NIterGrad  GSStotObjEval  timer (s)
prob9_1	      4	         45             1,973302126
prob9_2	      5	         67             1,865123034
prob10_1      3	         29             1,706352711
prob10_2      5	         41             2,131811142
probBrac      34	 490            14,33498287
probCart      4	         43             4,45336771
probZer       1	         0              4,276870251
probLand      40	 564            18,84315705

There are more changes that I want to do to the code, but I'll wait
until the next commit.