No more duplicates and better adjustments for trios
I have refactored all mentions to the trios, keeping the indexes to
the corresponding steps in the histStep attribute. Everything is
indexed right now. In order to do it, I had to include the "null
step" as the first tried step, which actually led to slightly better
performance.

I have also added checks in adjsTrios() to change the left and right
elements of the trios before trying new steps. I noticed that the
testDecr() method was broken (hopefully due to these changes), and
it was fixed.

By the way, I fell into a trap when testing with probRock because
the default case for probRock was not exactly the same as the one
for profiling (probRock-noPlot.its); there was a 1e-10 difference in
PFtol... I am solving this now as well.

There is not much more that can be done to increase the efficiency
of calcStepGrad. Maybe the proper implementation of the golden
section search, but I am really not sure.

These are the results for the batch test:

Problem name  NIterGrad  GSStotObjEval  timer (s)
prob9_1       4          44             1,973314047
prob9_2       5          70             1,735612154
prob10_1      3          28             1,58660078
prob10_2      5          41             2,148612261
probBrac      34         495            13,45475984
probCart      4          43             4,463190079
probZer       1          0              3,756063938
probLand      38         485            14,33326912