First advancements with cython - LMPBVP
I have "cythonized" the bottlenecks in LMPBVP, resulting in a time
reduction of 36% in LMPBVP, which gave a 30% reduction in the
"standard" test of 400 gradient iterations of probRock ("no plot").

There are still some changes to do in the code, but I am no longer
expecting a 10x time reduction. The future of this project gets less
and less promissing...