Divide and conquer
I am breaking the biggest villain of performance, c_profile, down to
three sub-functions, just for closest profiling. The intention is to
remove these subfunctions later as they actually slow down the pro-
gram.

The simp function was greatly improved, but that did not make that
much of a difference (hence, this dividing of the c_profile...)

"Standard comparison": pure python: 49.2 s; cython: 37.3 s.