Numpy is full of surprises!
After reading a lot of material about numpy and code vectorization,
I've decided to look up some tutorials in numpy's rather obscure
functions such as 'einsum'. It turns out that the computations that
take up most of the time in propagate() and prepMat() can be easily
replaced by judicious calls of einsum(), and sometimes swapaxes().

The result is that now the pure python running time was reduced al-
most by half! The cython running time is larger because I still have
to implement the same changes over there. I know, however, that its
time won't be so much lower as compared to pure python's, because
most of the for loops will be gone.

There are still a LOT of segments to be rewritten, using vectoriza-
tion guidelines and magic functions as einsum(). The future looks
promising again!

"Standard comparison": pure python: 26.8 s; cython: 33.2 s.