Failure while vectorizing 'Xi' propagation... improvements anyway
Yes, that did not work. It turns out that the cost of building the
big matrix to do the Xi propagation in one big step (and no 'for'
loops) is too great to offset the cost of the 'for' in time (N-1) in
each call of propagate().

Anyway, I continued the vectorization of the key parts of the code.
This time it was the turn for calcJ, the density function call in
probRock, and calcP.

For the next advancements, it will be interesting to modify the flow
of the program a little bit to avoid calculating things twice during
tryStep(), like the phi array, for example.

The main bottlenecks outside of einsum now are:
- the 'Xi' propagation, which I failed to improve
- the array declarations...?
- the density function, rhoSGRA and its auxiliary function, calSGRA
- the calcPhi function
- the calcDimCtrl function.

"Standard comparison": pure python: 23.5 s; cython: 31.6 s.