No more costly repeated calculations
I have come up with a way to avoid repeated calculations (e.g, the
phi and f arrays) that alters the code as little as possible, by the
use of a decorator. It turns out that the gains are not that expres-
sive, because of the looping scheme of the decorator, and having to
set all the 'isUpdated' flags to "False" when trying a new step. Be-
sides probRock, I also tested in prob9_1 and probLand; in these,
there was no significant change in running time.

The "standard comparison" test that I had been using has become
outdated because now most of the time is spend on disk-centered op-
erations, such as savefig(). Because of this, the running time
varies a lot.

This somewhat ends the cycle of optimizations, because now most of
the time is already been spent on c-level functions such as einsum.
I could still try to improve things like rhoSGRA(), but that answers
to ~5% of the running time in a long run (400 gradient iterations).

I will try to ask around for a way to further optimize propXi, be-
cause its share in a long run is much larger, ~19%.

"Standard comparison": pure python: 19.6 s; cython: 27.8 s.
(I swear I was able to get 19.4 s once, but no more. Who knows what
happened?)