Still fixing bugs in calcStepGrad...
Yes, I am still finding and fixing bugs in calcStepGrad. I have let
run a probRock instance with N = 201 (not 501), and a problem mani-
fested itself in iteration #2796. The good news is that it has been
fixed, and the case ran without problems, converging in iter. #3346.

These results reinforce the idea that lowering the value of N could
be the solution to yield reasonable running times. I know from pre-
vious experience, however, that the probLand requires a high N to
converge with acceleration limitation. ProbRock should be similar
in this aspect as well. I will now try to test the impact of N in
the accuracy of the solution in prob10_1 and 10_2, for which the
exact variational solution is known.