Working on low N validation in probLand
Continuing the process of trying to compare the MSGRA's final so-
lution and the "exact" variational solution for probLand, there
were problems when adapting from the "exact" solution, which comes
in very refined arrays, into the MSGRA framework, with two arcs
and a much less refined time array.

Just interpolating the arrays yields a solution that looks like
the exact solution (visually), but has a high P value. I have
spent a long time trying to interpolate the states and recalculate
the thrust so that the resulting P value is low, but that was shown
to be impossible.

After that was settled, I started to change the parameters, and
interesting results appeared. For documenting purposes, I am using
tolP = 1e-8, tolQ = 1e-4, acc_max = 5 g, acc_max_relTol = 0.1.
These parameters and all others were not changed as well.

With PFtol = 1e-2 and N = 101, MSGRA converges, with relative
error in I of 0.368% and RMS errors w.r.t. the exact variational
solution: x = 0.0272, u = 0.992, pi = 0.246.

With PFtol = 1e-1 and N = 101, MSGRA did not converge, seeming to
have hit a wall around GradIter #250. (Actually, the grad step
does a kind of a curve down, going below 1e-6, then up again above
1e-1, before hitting the wall. It could be interesting to investi-
gate this case further!). Increasing N to 1001, the calcStepGrad
algorithm broke down completely around GradIter #181 because
  23.4 = P(0.627) > P(1.) = 22.2 ,
a condition for which the code is not simply not prepared.

This calls for another refurbishment of calcStepGrad...