Another huge bug discovered in calcStepGrad...
Ok, maybe not a huge bug per se, but more of a huge vunerability of
the current algorithm. Let's tell the whole story:

First of all, the previous problem reported in last commit was
fixed by adding a new "bad point case", for when the step exceeds
some StepLimUppr, even if Obj<Obj0 and P<limP for this step value.

I also removed a few embarrassing mistakes here and there, such as a
rather bizarre procedure in raisObj() method in which step values
were being replaced by objective values... I also implemented
methods for returning the bestStep and bestObj so far, instead of
keeping copies of these values in the stepMngr object.

Anyway, the program still seemed to be "hitting the wall". This
time, things were getting weird around iteration 427. I noticed
that the I and J reduction "efficiencies" were quite low, even with
small values of step. (These quantities are calculated and displayed
as "I relative reduction", (I-I0)/alfa/(-self.Q), and "J relative re-
duction", (J-J0)/alfa/(-self.Q)). This really should not be the case,
because in principle

\delta I = \delta J = - \alpha * Q (eq 18a in Miele and Wang (2013)).

So, as alfa shrinks to 0, both (I-I0)/alfa/(-self.Q) and
(J-J0)/alfa/(-self.Q) should converge to 1 (100%).

I read through the log, and this is what happened:

ind  alfa        P              Irel.red   Jrel.red
0    0.0         1.768412E-11    0.0%       0.0%
1    1.8260E-04  1.802684E-11   -0.2%       0.5%
2    1.0         1.417096E+00   -5183.6%   -5183.6%
3    8.8574E-01  9.506391E-01   -4778.2%   -4778.2%
4    4.4296E-01  6.221073E-02   -2805.2%   -2805.2%
5    9.1301E-05  1.759973E-11   -0.5%      -0.5%

After attempt #5, the trio was set as [5,1,4], which does not make
any sense because step #5 is actually a bad step, but is smaller than
step #1, which is a good step. Of course the code crashed and burned
from this point onwards... maybe every case of wall hit is actually
one of these cases! I am commiting this run as well, under defaults/
probLand-treta.its.

This goes against a MAJOR assumption of the whole calcStepGrad algo-
rithm: "all steps smaller than a good step must also be good steps."

I have no choice but to re-design the entire part 2 of the step
search algorithm, or even maybe the whole algorithm itself... But
first, I am going to map the flow of calcStepGrad - part 2.