Towards serializing calcStepGrad tester
I have made a calcStepGrad tester, and it is somewhat serialized
already. One can create multiple dummy J vs. step and P vs. step be-
haviors, and test how current calcStepGrad algorithm runs on them.
I have not yet made many tests, but the calcStepGrad algorithm has
performed very well in all of the cases I've tested so far. Natural-
ly, things only get weird when the J and/or P functions are no longer
well-behaved. I was thinking about inserting some perturbations to
the J and P functions, in order to create fake local minima, but I
have another idea. It would be even better if I could "manually" in-
put some points for J and/or P and make a interpolation function out
of these points. With this, I would be able to test new algorithms
with the actual functions (well, not really, but close approximations
of them) that I will be dealing with in the real cases.

Let's go!