calcStepGrad tester OK with interpolation curves
As stated in the previous commit, I implemented the manual input
for points in J and P, so the sgraDummy object interpolates/ex-
trapolates a J or P vs. step behavior.

The next task is to make a parser that looks through a log file
from a MSGRA run and retrieves the steps that were evaluated, with
their corresponding values of P and J.