calcStepGrad tester: log parser OK
The log parser that goes through a given run folder of a previous
MSGRA execution is ready. It also gets parameters from the .its
file from the same folder, as to try to mimic the behavior of the
calcStepGrad algorithm in "real life".

I am now ready for designing a new calcStepGrad algorithm. It must
not only have good performance in simple problems, but also be ro-
bust enough in the perverse situations that some instances of
probLand (and most likely, probRock as well) present, e.g., more
than one local minimum for J.

Let's do it!