A new calcStepGrad algorithm
It is done. The new calcStepGrad algorithm is ready. It became
simpler and much more robust, at the cost of more objective eval-
uations.

The main idea is to declare an attempted step value 'a' as bad
(that is, invalid) if there is a smaller step 'b', whose objective
value is lower than a's. That is:

\exists b < a \mid Obj(b) < Obj(a) \implies a is invalid.

This means that when attempting some step value, other (larger)
previously tried steps can be relabeled as invalid. So, the algo-
rithm was modified to allow for "trio revolutions", where the trio
must be rebuilt from scratch.

It is easy to see that in "normal" situations, with only one local
minimum, nothing is lost and the algorithm works as before. If,
however, the program finds evidence of a lower local minimum, say,
by finding a positive gradient before the current local minimum,
it simply reassigns all steps after the positive gradient as in-
valid and starts over.

I also changed the P limit search, making it simpler and more ro-
bust also.

Unfortunately, my ideas in test-calcStepGrad did not work so well.
The current algorithm was tried with common sgra problems...

I have found some very interesting results indeed. But I will write
about them in the next commit only.