New algorithm, same results, great findings
After running the same case (defaults/probLand-treta.its), very
similar results were obtained. This time, the program was stop-
ped in iteration 405.

Later, talking with Muniz, he proposed that this issue should
also manifest itself in different problems. So, I decided to
perform a test in prob10-1 reducing the value of tolQ to see
at which point the program would stop converging. The results
were very interesting (source: defaults/exper-Q-10_1.bat):

n  tolQ  tolP  ok?  N  evals  I_rel,%  rms_x   rms_u   rms_pi
0  1E-4  1E-8  yes  2  17     8.1E-03  2.7E-3  3.5E-3  1.3E-4
1  1E-5  1E-8  yes  3  18.5   5.8E-03  2.0E-4  5.1E-4  9.0E-5
2  1E-6  1E-8  no   -  -      -        -       -       -
3  1E-6  1E-10 yes  4  10     8.2E-03  2.0E-4  2.1E-4  1.3E-4

In this table, "ok?" refers to whether the run was complete or
not, "evals" referes to the average number of objective evalua-
tions per gradient iteration, "I_rel,%" refers to the relative
error in I w.r.t. the exact minimum value of I and the last
columns correspond to the root-mean-square errors in x, u, and
pi, respectively, w.r.t. the exact variational solution.
The first row is associated with the "default" values of tolP
and tolQ suggested by Miele (1970).

If we tighten the Q tolerance to 1e-5, all the errors are re-
duced, as expected. However, if we reduce it even further to
1e-6, it simply does not converge anymore. Furthermore, the
steps that minimize the objective are:

GradIter  Step        IRed       IRedEff
1         3.9114E-1   9.7673E-3   49.97%
2         4.0516E-1   3.6306E-5  109.81%
3         4.8838E-14  0.0000E+0    0.00%

And from this point on, the step is lower and lower. Eventu-
ally, the step reaches zero, and the algorithm breaks down.
For the first iteration in which this behavior is observed,
the log of the gradient step search algorithm shows rapid va-
riations from large positive J reduction efficiencies to
large negative ones and vice-versa, even with step values as
low as 1e-11. There is only one hypothesis that supports this
effect, which is some sort of numerical noise. This means the
algorithm really cannot do any sensible improvement over the
running solution anymore.

The program could be brought to converge with tolQ=1e-6 by
reducing tolP to 1e-10. But then, another very interesting
finding is that the error in the final value of I w.r.t. the
actual exact variational solution is GREATER than the pre-
vious values of I, with tolQ=1e-4 AND 1e-5. This is huge!
It means that reducing tolQ alone is NOT SUFFICIENT to en-
sure that the solution is objectively closer to the exact
solution!

The very same behavior is observed in probLand, in the case
"defaults/probLand-treta.its". Here, the program seems to
run normally, but right after grad iteration 403, the same
"breakdown" occurred:

GradIter  Step         IRed       IRedEff
401       1.9918E-02   5.7787E-6   78.08%
402       9.6246E-02   9.1995E-6   45.78%
403       1.0471E-02   1.1524E-5  100.86%
404       4.3941E-14   0.0000E+0    0.00%
405       0.0000E+00   0.0000E+0    -----

And from this point onwards, the program broke completely.
Reading the log in the search of the grad step for iteration
number 403, the same behavior is observed, with rapid variations
from large positive J reduction efficiencies to large negative
ones and vice-versa, even with step values as low as 1e-12.

For probLand the implications are not good. The value of PFtol
simply cannot be increased as deemed necessary, because the pro-
blem is unable to converge to the same Q values (even if the
user increases N to ridiculous values like 10001). We can proba-
bly assume that the same observation holds for the rocket problem.

These results close the current endeavor of trying to re-obtain
Miele's results as of 2003. Besides all the other difficulties,
it is proven that there is no proper way of finding the proper
Kpf value.

The team is now deciding what are the next steps in this project.