Added "no extra arcs" possibility in probRock
I had decided to try this for a long time, but got caught up in
other ideas that I wanted to try first. It worked pretty well,
and made the program run much better. The same problem of "hit-
ting the wall" - which now is evidenced by the gradient step tur-
ning to zero - still manifests itselt, though.

So, I decided to add a "legitimate" exit condition in interf.py,
similar to what was done for too many gradient iterations (which,
by the way, is much less likely...). Doing this makes sure that all
the plots and .pkl's are saved in the end, which is desirable.

The next step will be (finally) adding a new mode to the batch
program (or alter the current 'variations' one) so that it is pos-
sible to use the final solution of a run as the initial guess of
the next one. This also should have been done a long time ago, but
now it will be useful for finding where exactly does the algorithm
stop working in terms of tolQ and Kpf.