No more starting over in batch!
I added a new "submode" to the batch program, enabling the use of
the final solution of a run as the initial guess of the other one.
It was easier than expected, honestly. There was a minor problem
with probLand because of the parameter loading methods
"loadParsFromFile" and "loadParsFromFile2" but it was solved.

Naturally, I went ahead and tested it for probLand (not probRock
yet), doing successive runs increasing Kpf (or rather PFtol) in-
crementally. The results were rather unexpected.

It turns out it is much worse for the algorithm to converge in a
problem that uses the acceleration limitation (with a non-negli-
gible Kpf value, of course) when the initial guess is a solution
that is already "converged" for a negligible Kpf. More accurate-
ly, the program does converge as Kpf is being gently increased,
but all of a sudden stops converging from a given value of Kpf.

Investigating this a little further, I discovered that this
highly sensitive behavior is due to the Q<tolQ comparison. When
Kpf is being gently raised, the Q values are also gently raised,
but still lower than tolQ, so the program exits without any
iterations (actually, 1 iteration because the initialization
counts as an iteration).

This problem will require a LOT of further investigation, so
I've decided to wrap this commit here.