Another brick in the wall-hitting problem
It turns out that the behavior previously mentioned was due to a
simple bug regarding avoidRepCalc. In the beginning of each run of
the batch, the value of Kpf was being changed, but our dear
avoidRepCalc was preventing the program from recalculating the ini-
tial value of I (which should change, but due to Kpf only).
This made the calcStepGrad program set itself to an impossible
task because I0 did not match I(0) (and neither did J0 and J(0)).

Fixing this, the batch specified in land-KpfSweep.bat as in the
previous commit worked (each one of its runs converged). I also
tried a simplified batch, consisting of the original
probLand-timer.its (PFtol=1e-6, tolQ=1e-3), followed by the
alterations in cascade mode:
1. PFtol=1e-3
2. PFtol=1e-2
3. PFtol=2e-2
each of these runs converges (until Q<tolQ, that is), with the end
of each run serving as the initial guess for the next one. However,
doing straight away PFtol=2e-2 makes the program hit a wall in it-
eration 66. The main conclusion here (which is also a somewhat sad
one), is that the initial guess makes a great difference in the
program's capacity for converging. The good news is that the pro-
gram does not seem to care if initialized with a "violating" solu-
tion. So this strategy of cascading solutions with progressively
higher values of Kpf may be viable for "real world" problems.

By the way, I also changed the runStatus reporting to stop calling
wall hits as 'successful runs'.

Experimenting a bit further, I deciced to vary other parameters,
such as N and the gradient step search's "P limit" constant,
abbreviated as GSS_PLimCte. As known, the first one is simply
the number of steps for the time discretization, and the latter
determines the maximum P value allowed. Increasing N tends to
allow greater values of Kpf, and decreasing GSS_PLimCte makes
the convergence process less aggressive, and hence more robust.
It turns out that increasing N from 101 to 201 makes the algo-
rithm converge (in 67 gradient iterations, which took 6s), from
the default initial guess. The program also converged from the
default initial guess with N = 101 by reducing GSS_PLimCte from
1e8 to 1e2 (in 534 gradient iterations, taking 16s this time).

So, the main lesson here is that increasing N and reducing the
GSS_PLimCte both improve the converging capacity of the program,
but with negative impacts on the time to converge. The trade-off
must be sought for each problem...

Going even deeper into the wall hit problem, I started to wonder
how can the program generate a correction that is simply unable
to lower the I or the J functionals? In principle, the first va-
riation of each of these functionals were supposed to be negati-
ve. I altered the program to plot the proposed variations in each
gradient step, and timestamp them to avoid overwriting. Running
probLand with this commit's probLand-alt.its, I could see the high
frequency oscillations on the correction induced by the sudden
shift in dynamics due to the acceleration limitation. I thought
that effect was to blame, but then I decided to try and tighten
tolQ for problem 10-1, to see what the corrections proposed by the
algorithm look like when the program is at its limit. For this
problem, the 'I' function is simply the time ('pi'). And indeed,
at gradient iteration 4, the program proposed a (tiny) correction
whose delta pi was 6.2e-7. This resulted in the same "impossible
task" mentioned before, because no gradient step could ever be va-
lid. This proves that the program is capable of proposing correc-
tions that it itself is unable to apply; resulting in a wall hit,
even without any penalty function. Furthermore, I even tried ini-
tializing prob10-1 and probCart with perfect (optimal) solutions.
In both cases, the program halted immediately due to very low
(albeit nonzero) Q values. Interestingly, though, the proposed
corrections were tiny but had positive values of delta 'pi'--they
would also result in wall hits if tolQ was tight enough. It is al-
so worth mentioning that besides positive delta 'pi's, these "im-
possible corrections" do not seem obvious results of numerical mis-
takes; they are smooth, well-behaved and look just as legitimate as
every other correction. This is curious because they are generated
by truncation errors and imperfect arithmetic during the solution
of the LSODE in "gradient mode".

It is hence established that the "wall hit" is a natural part of
the algorithm when it is pushed beyond its limits, even without a
penalty function. Naturally, the presence of the penalty function
greatly increases this possibility, by making the dependence of the
I (and J) functions with gradient step more complicated.